widen the encoding drag row from int to NSInteger
NSTableViewDataSource declares proposedRow: and row: as NSInteger, but
-validateDrop: and -acceptDrop: implemented them as int. Objective-C
dispatch matches on selector name alone, so the compiler could not catch
it: AppKit passed a 64-bit row in a register and the callee read the low
32 bits. Encoding rows never exceed eight, so the truncation was
harmless in practice -- but it was a signature mismatch waiting to bite
if either method were ever reused for a longer table.

Widening only the two parameters would have moved the truncation one
call deeper, so the whole path the row travels is now NSInteger: the
mEncodingDragRow ivar, the drag pasteboard payload, and
-moveEncodingFrom:to:.

Since the pasteboard payload changed width, -acceptDrop: now checks the
data length before decoding rather than dereferencing bytes as an int,
and -moveEncodingFrom:to: range-checks its indices. That method indexes
a fixed-size array and, for src == dst, would hand memmove a negative
length that converts to an enormous size_t. -validateDrop: rejects those
drops today, which is why this has never fired, but the guard belongs in
the method that would corrupt memory.

+++ cotvnc/Source/Profile.h
@@ -155,7 +155,7 @@ struct encoding {
 - (void)setTapAndClickTimeout:(NSTimeInterval)timeout
                     forButton:(unsigned)button;
 - (void)setEncodingEnabled:(BOOL)enabled atIndex:(int)index;
-- (void)moveEncodingFrom:(int)src to:(int)dst;
+- (void)moveEncodingFrom:(NSInteger)src to:(NSInteger)dst;
 - (void)setCopyRectEnabled:(BOOL)enabled;
 - (void)setJpegEncodingEnabled:(BOOL)enabled;
 - (void)setJpegLevel:(int)level;
