+++ cotvnc/Source/ProfileManager.h
@@ -65,7 +65,7 @@ typedef enum {
     IBOutlet NSButton *mEnableJpegEncoding;
     IBOutlet NSMatrix *mPixelFormatMatrix;
     IBOutlet NSColorWell *mTintColorWell;
-	int mEncodingDragRow;
+	NSInteger mEncodingDragRow;
 }
 
 	// Shared Instance
